refactor: 지도 부분 메서드 분리 및 타입 모듈화
