+++ src/routes/map/Map.tsx
@@ -1,187 +1,49 @@
-import Pin from "../../components/map/Pin";
 import React, { useEffect, useState, useRef } from "react";
-import Aim from '../../assets/image/icons/aim.png';
-import { createRoot } from "react-dom/client";
+import Aim from '@/assets/image/icons/aim.png';
+import { Property, OverlayData } from "@/utils/types";
+import { initializeMap } from "./methods/initializeMap";
+import { moveToCurrentLocation } from "./methods/moveToCurrentLocation";
+import { renderProperties } from "./methods/renderProperties";
+import { updateSelectedProperty } from "./methods/updateSelectedProperty";
 
-type Property = {
-  propertyId: number;
-  articleName: string;
-  latitude: number;
-  longitude: number;
-  buildingName: string;
-  realEstateType: string;
-  area1: number;
-  area2: number;
-  count: number;
-  price: number;
-};
 
 export default function MapComponent() {
   const [mapInfo, setMapInfo] = useState<string>('');
   const [properties, setProperties] = useState<Property[]>([]);
   const [selectedPropertyId, setSelectedPropertyId] = useState<number | null>(null);
   const [map, setMap] = useState<kakao.maps.Map|null>(null);
-  const overlayRef = useRef<{ [key: number]: any }>({});
+  const overlayRef = useRef<{ [key: number]: OverlayData }>({});
+  const previousSelectedPropertyIdRef = useRef<number | null>(null);
 
   // 더미데이터, 지도, 지도정보, 지도컨트롤러 생성
   useEffect(() => {
-    const dummyProperties: Property[] = [
-      {
-        propertyId: 1,
-        articleName: "102동 305호",
-        latitude: 37.54597011866081,
-        longitude: 127.0576325066247,
-        buildingName: "위례자이",
-        realEstateType: "아파트",
-        area1: 30,
-        area2: 27,
-        count: 5,
-        price: 500,
-      },
-      {
-        propertyId: 2,
-        articleName: "103동 307호",
-        latitude: 37.54697011866081,
-        longitude: 127.0586325066247,
-        buildingName: "위례자이",
-        realEstateType: "아파트",
-        area1: 28,
-        area2: 25,
-        count: 3,
-        price: 850,
-      },
-      {
-        propertyId: 3,
-        articleName: "104동 309호",
-        latitude: 37.547970118660814,
-        longitude: 127.0596325066247,
-        buildingName: "위례자이",
-        realEstateType: "아파트",
-        area1: 32,
-        area2: 29,
-        count: 1,
-        price: 1100,
-      },
-      {
-        propertyId: 4,
-        articleName: "104동 309호",
-        latitude: 37.53992802945424,
-        longitude: 127.05142662693137,
-        buildingName: "위례자이",
-        realEstateType: "아파트",
-        area1: 32,
-        area2: 29,
-        count: 1,
-        price: 1100,
-      },
-    ];
-
-    setProperties(dummyProperties);
-
-    const container = document.getElementById('map');
-    const options = {
-      center: new window.kakao.maps.LatLng(37.5449, 127.0566), // 지도의 중심좌표
-      level: 3, // 지도의 확대 레벨
-    };
-    if (!container){
-      return
-    }
-    const mapInstance = new window.kakao.maps.Map(container, options); // 지도를 생성
-    setMap(mapInstance);
-
-    const getInfo = () => {
-      const center = mapInstance.getCenter();
-      const level = mapInstance.getLevel();
-      const bounds = mapInstance.getBounds();
-      const swLatLng = bounds.getSouthWest();
-      const neLatLng = bounds.getNorthEast();
-
-      const latRange = swLatLng.getLat() + ' ~ ' + neLatLng.getLat();
-      const lngRange = swLatLng.getLng() + ' ~ ' + neLatLng.getLng();
-
-      let message = '지도 중심좌표는 위도 ' + center.getLat() + ', \n';
-      message += '경도 ' + center.getLng() + ' 이고 \n';
-      message += '지도 레벨은 ' + level + ' 입니다 \n\n';
-      message += '위도의 범위는 ' + latRange + ' 이고 \n';
-      message += '경도의 범위는 ' + lngRange + ' 입니다';
-
-      setMapInfo(message);
-    };
-
-    mapInstance.addControl(new window.kakao.maps.ZoomControl(), window.kakao.maps.ControlPosition.RIGHT);
-    mapInstance.addControl(new window.kakao.maps.MapTypeControl(), window.kakao.maps.ControlPosition.TOPRIGHT);
-    getInfo();
-
-    window.kakao.maps.event.addListener(mapInstance, 'center_changed', getInfo);
-    window.kakao.maps.event.addListener(mapInstance, 'zoom_changed', getInfo);
-
-    return () => {
-      window.kakao.maps.event.removeListener(mapInstance, 'center_changed', getInfo);
-      window.kakao.maps.event.removeListener(mapInstance, 'zoom_changed', getInfo);
-    };
+    const cleanup = initializeMap(setProperties, setMap, setMapInfo);
+    return cleanup;
   }, []);
 
-  // 현재 위치로
+  // 현재 위치로 아이콘 생성
   useEffect(() => {
-    if (map) {
-      const moveToCurrentLocation = () => {
-        if (navigator.geolocation) {
-          navigator.geolocation.getCurrentPosition((position) => {
-            const lat = position.coords.latitude;
-            const lon = position.coords.longitude;
-            const locPosition = new window.kakao.maps.LatLng(lat, lon);
-            map.panTo(locPosition);
-          });
-        }
-      };
-
-      document.getElementById('currentLocationImg')?.addEventListener('click', moveToCurrentLocation);
-
-      return () => {
-        document.getElementById('currentLocationImg')?.removeEventListener('click', moveToCurrentLocation);
-      };
-    }
+    const cleanup = moveToCurrentLocation(map);
+    return cleanup;
   }, [map]);
 
-  // 매물 그리기, return시 매물 지우기
+  // 전체 매물 그리기
   useEffect(() => {
-    if (!map) return;
-    
+    const cleanup = renderProperties(map, properties, overlayRef, selectedPropertyId, setSelectedPropertyId);
+    return cleanup;
+  }, [map, properties]);
 
-    properties.forEach((property) => {
-      const position = new window.kakao.maps.LatLng(property.latitude, property.longitude);
-      const container = document.createElement('div');
-      const root = createRoot(container); 
-      root.render(
-        <Pin
-          area={property.area1}
-          price={property.price}
-          count={property.count}
-          propertyId={property.propertyId}
-          handleClick={(id: number) => {
-            setSelectedPropertyId(id);
-          }}
-          isSelected={selectedPropertyId === property.propertyId}
-        />,
-      );
+  // 매물 선택시 스타일 변경
+  useEffect(() => {
+    updateSelectedProperty(
+      selectedPropertyId,
+      previousSelectedPropertyIdRef,
+      overlayRef,
+      properties,
+      setSelectedPropertyId
+    );
+  }, [selectedPropertyId, map, properties]);
 
-      const customOverlay = new window.kakao.maps.CustomOverlay({
-        position: position,
-        content: container,
-        yAnchor: 1,
-      });
-
-      customOverlay.setMap(map);
-
-      // 오버레이를 추적하여 이후에 제거할 수 있도록 함
-      overlayRef.current[property.propertyId] = customOverlay;
-    });
-
-    return () => {
-
-      Object.values(overlayRef.current).forEach((overlay) => overlay.setMap(null));
-    };
-  }, [map, properties, selectedPropertyId]);
 
   return (
     <div className="pt-16 h-[100vh]">
